Split trayicon.py into trayicon/ package
