+++ src/backend/trayicon/DBusService.py
@@ -389,64 +389,3 @@ class StatusNotifierItemService(DBusService):
             "XAyatanaNewLabel",
             (label, "")
         )
-
-class DBusTrayIcon:
-    def __init__(self, menu = None, path = "", menu_path = "", app_id = "", title = ""):
-        session_bus = Gio.bus_get_sync(Gio.BusType.SESSION)
-
-        self.menu = menu
-
-        kwargs = {
-            "session_bus": session_bus,
-            "menu_items": self.menu.get_items()
-        }
-        if path != "":
-            kwargs["path"] = path
-
-        if menu_path != "":
-            kwargs["menu_path"] = menu_path
-
-        self.sni_service = StatusNotifierItemService(**kwargs)
-        if app_id != "":
-            self.sni_service.Id = app_id
-
-        if title != "":
-            self.sni_service.Title = title
-
-    def set_icon(self, icon, path: str = ""):
-        self.sni_service.set_icon(icon, path)
-
-    def set_tooltip(self, title, description = ""):
-        self.sni_service.set_tooltip(title, description)
-
-    def set_label(self, label):
-        self.sni_service.set_xayatanalabel(label)
-
-    def update_menu(self):
-        self.sni_service.set_items(self.menu.get_items())
-
-    def register(self):
-        self.sni_service.register()
-
-    def unregister(self):
-        self.sni_service.unregister()
-
-class DBusMenu:
-    def __init__(self):
-        self.menu_items = []
-
-    def add_menu_item(self, menu_id, menu_label="", menu_type="", icon_name="", callback=None):
-        item = {'id': menu_id}
-        if menu_label != "":
-            item['label'] = menu_label
-        if menu_type != "":
-            item['type'] = menu_type
-        if icon_name != "":
-            item['icon-name'] = icon_name
-        if callback:
-            item['callback'] = callback
-
-        self.menu_items.append(item)
-
-    def get_items(self):
-        return self.menu_items
