+++ src/backend/trayicon/__init__.py
@@ -1,0 +1,19 @@
+from src.backend.trayicon.DBusMenu import DBusMenu
+from src.backend.trayicon.DBusService import (
+    DBusMenuService,
+    DBusService,
+    MENU_NODE_INFO,
+    SNI_NODE_INFO,
+    StatusNotifierItemService,
+)
+from src.backend.trayicon.DBusTrayIcon import DBusTrayIcon
+
+__all__ = [
+    "DBusMenu",
+    "DBusMenuService",
+    "DBusService",
+    "DBusTrayIcon",
+    "MENU_NODE_INFO",
+    "SNI_NODE_INFO",
+    "StatusNotifierItemService",
+]
